Refactor Admin from interface to class
Changed the structure of the model package a bit regarding admins and employees.
Made Admin a class that extends employee and not an interface.
Employee class now has protected setters so that admins can easily set their SIN
and citizenship status. Employees can also adjust their SIN in theory, but if we
are careful of how we implement the functionality in the GUI, then only admins will
be able to change employee information.